requirements.txt
Flask==2.3.2
python-docx==0.8.11
reportlab==3.6.12
